requirements.txt
sentence-transformers
transformers
torch
faiss-cpu
pymupdf
numpy
accelerate